Move the ActiveRecord extensions into their own file so they can be required separately
